refactor: 마이냠 관련 페이지들을 /screen/my_page 디렉토리로 이동
