+++ lib/screen/my_page/favorite_stores_screen.dart
